Optimize `AABBMesh.cpp` squared_distance() methods.
For some reason there was an excess of vector and index argument copying up
and down the various layers of implementation that I've stripped out. I also
changed some methods to use the standard size_t as the indexed_triange_set
index type instead of int because it makes the method calls pass straight
through the implementation layers.

+++ src/slic3r/GUI/MeshUtils.cpp
@@ -564,7 +564,7 @@ bool MeshRaycaster::closest_hit(const Vec2d& mouse_pos, const Transform3d& trafo
 
 Vec3f MeshRaycaster::get_closest_point(const Vec3f& point, Vec3f* normal) const
 {
-    int idx = 0;
+    size_t idx = 0;
     Vec3d closest_point;
     Vec3d pointd = point.cast<double>();
     m_emesh.squared_distance(pointd, idx, closest_point);
@@ -577,7 +577,7 @@ Vec3f MeshRaycaster::get_closest_point(const Vec3f& point, Vec3f* normal) const
 
 int MeshRaycaster::get_closest_facet(const Vec3f &point) const
 {
-    int   facet_idx = 0;
+    size_t facet_idx = 0;
     Vec3d closest_point;
     m_emesh.squared_distance(point.cast<double>(), facet_idx, closest_point);
     return facet_idx;
